Organised files, added notebooks, using new basis functions
Organised the files of the github to separate the notebooks from the different KAN variants (taken from their respective git repositories), and other code that I have written.

Added notebooks that show the training of the chebyshev kan basis function on the Lotka-Volterra and duffing equations. Different initial conditions for the duffing oscillator have also been investigated although the training gets caught in local minima most of the time

Created a class in the ChebyKANLayer.py file to automatically initialise cKANs to avoid manually
setting layers, the format is consistent with efficientKAN.