Testing FastKAN and creating a class to train models
I ran one test on the FastKAN package becuase it uses the GRBF as a basis function for its KAN implementation.

Working on a class that should be able to train and test models modularly, this avoids copying and pasting the same code everytime I want to change something about how the model is trained - incomplete.