Added the functionality to save and load models to the ChebyKAN package
Implemented functions to load and save models to the ChebyKAN class. It works however some flaws to look out for is that it assumes the class that you have initialised has the same size and amount of layers as the model you are trying to load.

It also creates a directory if one doesn't exist where you are trying to save