Weekly Work
Refactored one of the core parameters of the main class, the samples input is now samples per second. This was to make it more convenient when setting the training times, and in my opinion more intuitive.

Ran the code for extended prediction times

Ran the code with different initial conditions for an already trained model

Took a sweep over the training time parameter (several times) for a model learning the LV equations

Tried to create another animation to visualise the evolution of coefficients, ran into kernel crashes which I now know are from running out of memory (due to the refactored parameter mentioned above)

Ran the program on the van der poll equation to test its capability in modelling limit cycles