week 7 work
Reverted changes to the main class regarding the usage of the GPU because it was slower.
Refactored the variable responsible for determining the amount of steps the integrator takes.  steps per second --> step size
Started the parameter sweep over the order of the basis function. Currently on iteration 7/10, optimal seems to be 4 which is what I've been using thus far from the paper that inspired this project.